+++ media/main.js
@@ -66,15 +66,16 @@ function handleMessage(message) {
       webviewDebugLogger(`Received prs: ${JSON.stringify(message.data)}`);
       addPRsToDoc(message.data.sortedPRs);
       // jira
-      if (message.data?.mostRelevantJiraTicket) {
+      if (message.data?.mostRelevantJiraTickets) {
         $("#mostRelevantJiraTicketHolder").empty();
-        addMostRelevantJiraTicket(message.data.mostRelevantJiraTicket);
+        addMostRelevantJiraTicket(message.data.mostRelevantJiraTickets);
       }
       clampCodeBlocks();
       break;
     case "error":
       webviewDebugLogger(`Received error: ${JSON.stringify(message.data)}`);
-      errorTimeout = setReceivedError(message.error.errorText, errorTimeout);
+      clearTimeout(errorTimeout);
+      setReceivedError(message.error.errorText);
       break;
     case "versionInfo":
       webviewDebugLogger(