+++ media/utils/DailySummary/Jira/addJiraTicketsToDailySummary.js
@@ -3,10 +3,24 @@ const addJiraTicketsToDailySummary = (jiraTickets) => {
   if (Array.isArray(jiraTickets)) {
     jiraTickets?.forEach((ticket) => {
       $("#dailySummary").append(`
-    <div class="Box-header">
-        <h5 class="Box-title Truncate">${ticket?.key}</h5>
-        <span style="color: red">${ticket?.fields?.status?.name}</span>
-        <p>${ticket?.fields?.summary}</p>
+      <div class="Box">
+      <div class="Box-header d-flex flex-justify-between">
+      <a href="${ticket.fields.priority.iconUrl.split("/images")[0]}/browse/${
+        ticket?.key
+      }">
+        <h5 class="Box-title Truncate">
+         ${ticket?.key}: ${ticket?.fields?.summary}
+       </h5>
+      </a>
+      <span style="color: green">${ticket?.fields?.status?.name}</span>
+    </div>
+    ${
+      ticket?.renderedFields?.description ? (
+        <div class="Box-body">${ticket?.renderedFields?.description}</div>
+      ) : (
+        ""
+      )
+    }
     </div>
     <br/>
     `);